Moved key bindings to plugin configurations
Keybindings were originally added to the mappings.lua file for all
plugins. This can be problematic, because there is no tight coupling
between keybindings that are newly introduced by plugins and the plugins
themselves.

This PR changes the code so that each plugin configuration also
introduces the keybinds associated with it. This way, if the plugin is
removed or changed in any way in the future, the developer does not have
to worry about also updating the mappings file manually.