Added proper support for Golang development
Added and configured a number of plugins for proper development in
Golang. These plugins include null-ls, lspconfig, gopher.nvim, and
dap plugins for debugging. In addition to these changes related to Go,
there is an initial installer script added which doesn't do anything
now, but should be written so that all prerequisites for running neovim
on a Linux system are installed upon running it.