Fixed error message neovim would print when running LspInfo command
When running :LspInfo, Neovim would print an error message stating that
an exception was occurring due to a bad argument, where a table was
expected and instead a string was recieved. The issue was that I was
indeed passing a variable as a string instead of a one item table in my
Lspconfig configuration. This commit fixes this by replacing 'python'
with { 'python' }.

In addition to this, the README has been updated with a little bit of
content.